% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBinByCellMatrix.R
\name{getBinByCellMatrix}
\alias{getBinByCellMatrix}
\title{Get the BinByCell matrix}
\usage{
getBinByCellMatrix(
  frag,
  genes,
  CpG,
  exon,
  Blacklist,
  tileSize = 500,
  excludeChr = c("chrY", "chrM")
)
}
\arguments{
\item{frag}{A \code{GRanges} object that contains fragment information.}

\item{genes}{A \code{GRanges} object that contains genes information.}

\item{CpG}{A \code{GRanges} object that contains CpG information.}

\item{exon}{A \code{GRanges} object that contains exon information.}

\item{Blacklist}{A \code{GRanges} object that contains Blacklist information.}

\item{tileSize}{The size of the tiles used for binning counts in the BinByCellMatrix.}

\item{excludeChr}{A character vector containing the \code{seqnames} of the chromosomes that should be excluded from the BinByCellMatrix.}
}
\value{
BinByCellMatrix
}
\description{
Get the BinByCell matrix
}
