% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneByGroupMatrix.R
\name{getGeneByGroupMatrix}
\alias{getGeneByGroupMatrix}
\title{Title}
\usage{
getGeneByGroupMatrix(
  genes,
  binByGroup,
  excludeChr = c("chrY", "chrM"),
  useGeneBoundaries = TRUE,
  extendUpstream = 2000,
  extendDownstream = 2000
)
}
\arguments{
\item{genes}{A \code{GRanges} object that contains genes information.}

\item{binByGroup}{binByGroup Matrix}

\item{excludeChr}{A character vector containing the \code{seqnames} of the chromosomes that should be excluded from the BinByCellMatrix.}

\item{useGeneBoundaries}{A boolean value indicating whether gene boundaries should be employed during gene activity score}

\item{extendUpstream}{The minimum and maximum number of basepairs upstream of the transcription start site to consider for gene
activity score calculation.}

\item{extendDownstream}{The minimum and maximum number of basepairs downstream of the transcription termination site to consider for
gene activity score calculation.}
}
\value{
GeneByGroup
}
\description{
Title
}
