% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHVbinMatrix.R
\name{getHVbinMatrix}
\alias{getHVbinMatrix}
\title{Get the High variability bin matrix}
\usage{
getHVbinMatrix(binByGroup_norm, bin_max_thr = 7, bin_fc_thr = 70)
}
\arguments{
\item{binByGroup_norm}{binByGroup_norm matrix}

\item{bin_max_thr}{Used to sieve out very low peaks, i.e. peaks with small maximum values, which may be noise.}

\item{bin_fc_thr}{Used to screen the bin for high variance}
}
\value{
binByGroup_norm_HVbin
}
\description{
Get the High variability bin matrix
}
